add PaymentDetailsController and paymentDetailsListFragment(paymentDetailsList).html

+++ autoDeal/src/main/java/com/example/autodeal/user/service/UserService.java
@@ -68,9 +68,12 @@ public class UserService implements UserDetailsService {
     }
 
     public void deleteUser(Integer id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User with ID " + id + " does not exist");
+        }
+
         userRepository.deleteById(id);
     }
-
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         UserModel user = findUserByEmail(email);
